requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-pptx==1.0.2
PyMuPDF==1.24.10
httpx==0.27.2
python-dotenv==1.0.1
